Extract common dashboard components to wrapper
