+++ src/components/DashboardWrapper.tsx
@@ -9,43 +9,14 @@ import {
   VStack,
 } from '@chakra-ui/layout';
 import { useBreakpointValue } from '@chakra-ui/media-query';
-import GradientButton from '../components/GradientButton';
-import StyledLink from '../components/StyledLink';
+import GradientButton from './GradientButton';
+import StyledLink from './StyledLink';
 import useAuth from '../hooks/useAuth';
 import { BACKGROUND_GRADIENT, BACKGROUND_GREY } from '../themes/colors';
-import Splash from '../components/Splash';
+import Splash from './Splash';
 import Link from 'next/link';
-import DashboardCard from '../components/DashboardCard';
 
-const NAVIGATION = [
-  {
-    title: 'Create new member/pairing',
-    description: 'Add new entry to database',
-    href: '/dashboard',
-  },
-  {
-    title: 'View members',
-    description: 'View all registered PSA members by class',
-    href: '/dashboard',
-  },
-  {
-    title: 'View pairings',
-    description: 'View all registered AKA pairings by semester',
-    href: '/dashboard',
-  },
-  {
-    title: 'View lineages',
-    description: 'Browse through AKA family trees',
-    href: '/dashboard',
-  },
-  {
-    title: 'View issues',
-    description: 'View pending administrative changes to be made',
-    href: '/dashboard',
-  },
-];
-
-const DashboardPage = () => {
+const DashboardWrapper = ({ children }: { children: React.ReactNode }) => {
   const { user, loadSplash } = useAuth();
   const isMobile = useBreakpointValue({ base: true, md: false });
 
@@ -110,26 +81,11 @@ const DashboardPage = () => {
               </Link>
             </Flex>
           )}
-          <Heading mb="3rem">
-            <span style={{ fontWeight: 'lighter' }}>Hello,</span>{' '}
-            {userFirstName}
-          </Heading>
-          <Grid
-            templateColumns={{ base: '100%', md: 'repeat(3, 1fr)' }}
-            gap={6}
-          >
-            {NAVIGATION.map(({ title, description, href }) => (
-              <DashboardCard
-                title={title}
-                description={description}
-                href={href}
-              />
-            ))}
-          </Grid>
+          {children}
         </Box>
       </Grid>
     </Box>
   );
 };
 
-export default DashboardPage;
+export default DashboardWrapper;
